feat(#140): replace component
