+++ packages/epics/src/agreements/components/agreement-card.tsx
@@ -1,5 +1,5 @@
 import { Text } from '@radix-ui/themes';
-import { Card, BadgeCva, Skeleton, StatusBadge } from '@hypha-platform/ui';
+import { Card, Badge, Skeleton, StatusBadge } from '@hypha-platform/ui';
 import { Image } from '@hypha-platform/ui';
 import { EyeOpenIcon, ChatBubbleIcon } from '@radix-ui/react-icons';
 
@@ -52,27 +52,27 @@ export const AgreementCard: React.FC<AgreementCardProps> = ({
       <div className="flex justify-between items-center w-full">
         <div className="flex flex-col">
           <div className="flex gap-x-1">
-            <BadgeCva
+            <Badge
               variant="surface"
               colorVariant="accent"
               isLoading={isLoading}
             >
               Agreement
-            </BadgeCva>
-            <BadgeCva
+            </Badge>
+            <Badge
               variant="surface"
               colorVariant="accent"
               isLoading={isLoading}
             >
               Recurring
-            </BadgeCva>
-            <BadgeCva
+            </Badge>
+            <Badge
               variant="surface"
               colorVariant="accent"
               isLoading={isLoading}
             >
               {commitment}%
-            </BadgeCva>
+            </Badge>
             <StatusBadge status={status} isLoading={isLoading} />
           </div>
