+++ packages/epics/src/treasury/components/requests/request-card.tsx
@@ -1,5 +1,5 @@
 import { Text } from '@radix-ui/themes';
-import { Card, BadgeCva } from '@hypha-platform/ui';
+import { Card, Badge } from '@hypha-platform/ui';
 import Image from 'next/image';
 import { CalendarIcon } from '@radix-ui/react-icons';
 import { formatDate } from '@hypha-platform/ui-utils';
@@ -44,20 +44,16 @@ export const RequestCard: React.FC<RequestCardProps> = ({
       <div className="flex justify-between items-center w-full">
         <div className="flex flex-col">
           <div className="flex gap-x-1">
-            <BadgeCva
+            <Badge
               isLoading={isLoading}
               variant="surface"
               colorVariant="accent"
             >
               {symbol}
-            </BadgeCva>
-            <BadgeCva
-              isLoading={isLoading}
-              variant="surface"
-              colorVariant="warn"
-            >
+            </Badge>
+            <Badge isLoading={isLoading} variant="surface" colorVariant="warn">
               Pending
-            </BadgeCva>
+            </Badge>
           </div>
           <Amount isLoading={isLoading} value={value} withUsdSymbol />
           <Skeleton loading={isLoading} width="80px" height="16px">
